requirements.txt
accelerate==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
causal-conv1d==1.6.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cuda-toolkit==12.8.1
einops==0.8.2
filelock==3.25.2
fsspec==2026.2.0
gitdb==4.0.12
gitpython==3.1.47
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.11.0
idna==3.4
jinja2==3.1.6
lightning-utilities==0.15.3
mamba-ssm==2.3.1
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.19.0.56
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==24.1
pillow==12.1.1
platformdirs==4.9.6
protobuf==7.34.1
psutil==7.2.2
pydantic==2.13.3
pydantic-core==2.46.3
pygments==2.20.0
pyyaml==6.0.3
regex==2026.4.4
requests==2.28.1
rich==15.0.0
safetensors==0.7.0
sentry-sdk==2.58.0
setuptools==70.2.0
shellingham==1.5.4
smmap==5.0.3
sympy==1.14.0
tokenizers==0.22.2
torch==2.11.0+cu128
torchmetrics==1.9.0
torchvision==0.26.0+cu128
tqdm==4.66.5
transformers==5.6.0
triton==3.6.0
typer==0.24.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==1.26.13
wandb==0.26.1
